Fix ARM64 compiler warnings in Finder and core modules
- Fix pointer-to-int-cast warnings with (uintptr_t) intermediate cast
- Fix shadow warning by renaming isTextFile to isTextDoc in folder_window.c
- Fix unused variable warnings with (void) casts for debug-only variables
- Fix cast-qual warnings with const-correct pointer casts
- Fix type-limits warning using INT16_MAX instead of UINT16_MAX for short

+++ src/Finder/AboutThisMac.c
@@ -449,7 +449,7 @@ void AboutWindow_ProcessPendingCreation(void)
     SelectWindow(sAboutWin);
 
     /* Request update */
-    PostEvent(updateEvt, (UInt32)sAboutWin);
+    PostEvent(updateEvt, (UInt32)(uintptr_t)sAboutWin);
 
     serial_puts("[ABOUT] ProcessPending: Window shown successfully\n");
 }
@@ -471,7 +471,7 @@ void AboutWindow_ShowOrToggle(void)
         serial_puts("[ABOUT] ShowOrToggle: Window exists, bringing to front\n");
         BringToFront(sAboutWin);
         SelectWindow(sAboutWin);
-        PostEvent(updateEvt, (UInt32)sAboutWin);
+        PostEvent(updateEvt, (UInt32)(uintptr_t)sAboutWin);
         FINDER_LOG_DEBUG("AboutThisMac: Window already shown\n");
         return;
     }
